Corrected MTF header size for >16bit compilers.

+++ mtf.h
@@ -59,8 +59,8 @@ typedef char INT8;
 typedef unsigned char UINT8;
 typedef short INT16;
 typedef unsigned short UINT16;
-typedef long INT32;
-typedef unsigned long UINT32;
+typedef int INT32;
+typedef unsigned int UINT32;
 
 typedef struct
 {
